+++ src/main/java/proyecto/huellitas/demo/servicio/MascotaService.java
@@ -5,11 +5,11 @@ import proyecto.huellitas.demo.entidad.Mascota;
 
 public interface MascotaService {
 
-    public Mascota SearchById(int id);
+    public Mascota SearchById(Long id);
 
     public Collection<Mascota> SearchAll();
 
-    public void DeleteById(int id);
+    public void DeleteById(Long id);
 
     public void Update(Mascota mascota);
